refactor(app): split generate screen modules
